add header authorization for ajax

+++ CinemaManagement/DataAccess/DAOs/UserDAO.cs
@@ -94,7 +94,7 @@ namespace DataAccess.DAOs
 			}
 		}
 
-        public List<Booking> GetAllSeatBookedByUserId(int id)
+        public List<Booking> GetAllSeatBookedByUserId(long id)
         {
             return new CinemaContext().Booking.Include(x => x.Show).Include(x => x.Show.Film).Where(x => x.UserId == id).ToList();
         }
